builds + passes tests
