+++ src/passes/index.ts
@@ -1,0 +1,3 @@
+export { tryTraverseSerializeParameter } from "./serializerParameter";
+export { tryTraverseComment } from "./comment";
+export * from "./types";
