+++ src/passes/visitNode/visitLetExpression.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import * as PQP from "@microsoft/powerquery-parser";
-import { SerializeParameterState, SerializerWriteKind } from "./../types";
+import { SerializeParameterState, SerializerWriteKind } from "../types";
 import { propagateWriteKind, setWorkspace } from "./visitNodeUtils";
 
 export function visitLetExpression(state: SerializeParameterState, node: PQP.Language.Ast.LetExpression): void {
