+++ src/passes/visitNode/visitNode.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import * as PQP from "@microsoft/powerquery-parser";
-import { SerializeParameter, SerializeParameterState } from "./../types";
+import { SerializeParameter, SerializeParameterState } from "../types";
 import { getWorkspace, propagateWriteKind, setWorkspace } from "./visitNodeUtils";
 
 import { visitArrayWrapper } from "./visitArrayWrapper";
