+++ src/passes/visitNode/visitSection.ts
