+++ src/passes/visitNode/visitSectionMember.ts
