+++ src/passes/visitNode/visitTPairedConstant.ts
