+++ src/passes/visitNode/visitTWrapped.ts
