+++ src/passes/visitNode/visitTableType.ts
