--- src/serializeParameter/isMultiline/common.ts
@@ -1,23 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-import * as PQP from "@microsoft/powerquery-parser";
-import { IsMultilineMap } from "../types";
-
-export function expectGetIsMultiline(isMultilineMap: IsMultilineMap, node: PQP.Language.Ast.TNode): boolean {
-    const maybeIsMultiline: boolean | undefined = isMultilineMap.get(node.id);
-    if (maybeIsMultiline === undefined) {
-        const details: {} = { nodeId: node.id };
-        throw new PQP.CommonError.InvariantError(`isMultiline is missing an expected nodeId`, details);
-    }
-
-    return maybeIsMultiline;
-}
-
-export function setIsMultiline(
-    isMultilineMap: IsMultilineMap,
-    node: PQP.Language.Ast.TNode,
-    isMultiline: boolean,
-): void {
-    isMultilineMap.set(node.id, isMultiline);
-}
